Refactored the project structure. Added apikey.properties file to manage API key
