+++ .gitignore
@@ -20,6 +20,9 @@ build/
 # Local configuration file (sdk path, etc)
 local.properties
 
+# File containing the API key(s)
+apikey.properties
+
 # Proguard folder generated by Eclipse
 proguard/
