+++ app/build.gradle.kts
@@ -1,4 +1,6 @@
 import org.jetbrains.kotlin.gradle.dsl.KotlinJvmOptions
+import java.util.Properties
+import java.io.FileInputStream
 
 plugins {
     id("com.android.application")
@@ -8,6 +10,10 @@ plugins {
 
     id("jacoco-android")
 }
+
+val apikeyPropertiesFile = rootProject.file("apikey.properties")
+val apikeyProperties = Properties()
+apikeyProperties.load(FileInputStream(apikeyPropertiesFile))
 
 android {
     compileSdkVersion(29)
@@ -22,16 +28,13 @@ android {
 
     signingConfigs {
         getByName("debug") {
-            keyAlias = project.property("keyAlias") as String
-            keyPassword = project.property("keyPassword") as String
-            storeFile = rootProject.file(project.property("storeFile") as String)
-            storePassword = project.property("storePassword") as String
+
         }
     }
 
     buildTypes {
         getByName("debug") {
-            buildConfigField("String", "API_KEY", project.property("MOVIE_DB_API_KEY") as String)
+            buildConfigField("String", "API_KEY", apikeyProperties["MOVIE_DB_API_KEY"] as String)
             signingConfig = signingConfigs.getByName("debug")
         }
     }
