+++ app/src/main/java/com/hitanshudhawan/popcorn2/PopCorn2Application.kt
@@ -1,6 +1,7 @@
 package com.hitanshudhawan.popcorn2
 
 import android.app.Application
+import com.hitanshudhawan.popcorn2.koin.networkKoinModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
 
@@ -13,7 +14,7 @@ class PopCorn2Application : Application() {
             androidContext(this@PopCorn2Application)
             modules(
                 listOf(
-                    networkKoinModule
+					networkKoinModule
                 )
             )
         }
