+++ app/src/main/java/com/hitanshudhawan/popcorn2/fragments/TVShowsFragment.kt
@@ -1,15 +1,16 @@
-package com.hitanshudhawan.popcorn2
+package com.hitanshudhawan.popcorn2.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.hitanshudhawan.popcorn2.R
 
-class FavoritesFragment : Fragment() {
+class TVShowsFragment : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.fragment_favorites, container, false)
+        return inflater.inflate(R.layout.fragment_tv_shows, container, false)
     }
 
 }
